Remove WorldCoinSlot, rewire start logic, and add CoinSlot to Cabinet class
