--- src/components/CoinSlot/styles.tsx
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const CanvasParent = styled.div`
-  position: relative;
-  width: 100%;
-  height: 100%;
-
-  z-index: 0;
-`;
