+++ src/webgl/config/buttons.ts
@@ -2,7 +2,7 @@ export const BUTTONS = [
   {
     key_value: "1",
     position: {
-      x: 1.14,
+      x: 1.04,
       y: 1.3,
       z: 2.85,
     },
@@ -10,7 +10,7 @@ export const BUTTONS = [
   {
     key_value: "2",
     position: {
-      x: 1.625,
+      x: 1.525,
       y: 1.3,
       z: 2.85,
     },
@@ -18,7 +18,7 @@ export const BUTTONS = [
   {
     key_value: "3",
     position: {
-      x: 2.1,
+      x: 2.0,
       y: 1.3,
       z: 2.85,
     },
@@ -26,7 +26,7 @@ export const BUTTONS = [
   {
     key_value: "4",
     position: {
-      x: 1.14,
+      x: 1.04,
       y: 1,
       z: 2.85,
     },
@@ -34,7 +34,7 @@ export const BUTTONS = [
   {
     key_value: "5",
     position: {
-      x: 1.625,
+      x: 1.525,
       y: 1,
       z: 2.85,
     },
@@ -42,7 +42,7 @@ export const BUTTONS = [
   {
     key_value: "6",
     position: {
-      x: 2.1,
+      x: 2.0,
       y: 1,
       z: 2.85,
     },
@@ -50,7 +50,7 @@ export const BUTTONS = [
   {
     key_value: "7",
     position: {
-      x: 1.14,
+      x: 1.04,
       y: 0.7,
       z: 2.85,
     },
@@ -58,7 +58,7 @@ export const BUTTONS = [
   {
     key_value: "8",
     position: {
-      x: 1.625,
+      x: 1.525,
       y: 0.7,
       z: 2.85,
     },
@@ -66,7 +66,7 @@ export const BUTTONS = [
   {
     key_value: "9",
     position: {
-      x: 2.1,
+      x: 2.0,
       y: 0.7,
       z: 2.85,
     },
@@ -74,7 +74,7 @@ export const BUTTONS = [
   {
     key_value: "C",
     position: {
-      x: 1.14,
+      x: 1.04,
       y: 0.4,
       z: 2.85,
     },
@@ -82,7 +82,7 @@ export const BUTTONS = [
   {
     key_value: "0",
     position: {
-      x: 1.625,
+      x: 1.525,
       y: 0.4,
       z: 2.85,
     },
@@ -90,7 +90,7 @@ export const BUTTONS = [
   {
     key_value: "E",
     position: {
-      x: 2.1,
+      x: 2.0,
       y: 0.4,
       z: 2.85,
     },
